Add type.bool
